👔 커뮤니티 인기 포스팅 컴포넌트 분리 ✨ 메인 화면 커뮤니티 인기 포스팅 출력 구현
